+++ sidebars.js
@@ -63,6 +63,7 @@ const sidebars = {
     "how-to-guides/poh-questn",
     "how-to-guides/poh-zealy",
     "how-to-guides/poh-galxe",
+    "how-to-guides/proof-of-humanity-collabland-discord",
     "how-to-guides/polygon-id-state-replication",
   ],
   reference: [