+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Use-cases:
+  - Added the "Rarimo x Identity State Replication" documentation page
+
 ## [1.6.0] - 28-11-2023
 
 ### Added