+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Reference:
+  - Added the "Polygon ID State Replication" page;
 - Use Cases:
   - Added the "Identity State Replication" doc;
 - How-to-Guides: